backend-mvc/requirements.txt
Flask
requests
flask_cors
pymongo
flask_jwt_extended
pyjwt
argon2-cffi
mongoengine